+++ src/check.c
@@ -9,6 +9,7 @@
 SEXP CCondnCheck(SEXP X){
   int p,i,j;  
   double *xptr;
+  double avg;
   xptr = REAL(X);
   p = INTEGER(GET_DIM(X))[0];
 
@@ -18,9 +19,10 @@ SEXP CCondnCheck(SEXP X){
 
   for ( i = 0; i < p; i++ ) {
     for ( j = 0; j <= i ; j++ ) {
-      if (xptr[i + j*p] > 0.5*(xptr[i + i*p] + xptr[j + j*p])) {
+      avg = 0.5*(xptr[i + i*p] + xptr[j + j*p]);
+      if ((xptr[i + j*p] > avg) || (xptr[j + i*p] > avg)) {
         LOGICAL(out)[0] = FALSE;
-        j = p + 1; //to break outer for loop as well
+        i = p + 1; //to break outer for loop as well
         break; //breaks inner for loop
       }
     }
@@ -29,4 +31,3 @@ SEXP CCondnCheck(SEXP X){
   UNPROTECT(1);
   return(out);
 }
-